pyproject.toml
[tool.poetry]
name = "lignin"
version = "0.1.0"
description = ""
authors = ["FoxCheeze <iagocarvalhodm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
mypy = "^0.971"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
